--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "react-pomodoro-app"
-  }
-}
