+++ package.json
@@ -2,6 +2,7 @@
   "name": "pomodoro-app",
   "version": "1.0.0",
   "private": true,
+  "homepage": "http://lexiewh.github.io/react-pomodoro-app",
   "dependencies": {
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.7",
