+++ src/components/Form/TimeForm.js
@@ -16,9 +16,8 @@ function TimeForm() {
   } = useForm(submit, validate)
 
   function submit() {
-    sessionStorage.setItem("times", JSON.stringify(values))
-    console.log(JSON.parse(sessionStorage.getItem('times')).prodTime)
-    history.push(`/timer`)
+    console.log(values)
+    history.push(`/timer/${values.prodTime}-${values.shortBreak}-${values.longBreak}-${values.interval}`)
   }
 
   return(
