+++ src/index.js
@@ -4,11 +4,6 @@ import './stylesheets/main.css'
 import 'semantic-ui-css/semantic.min.css'
 
 // Components
-import App from './App';
+import App from './components/App';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+ReactDOM.render(<App />, document.getElementById('root'));
